requirements.txt
numpy==1.15.4
scikit-image==0.14.1
pyqt5==5.11.3
flask==1.0.2
pymongo==3.7.2
attrs==19.1.0